+++ CHANGELOG.md
@@ -20,15 +20,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Improve settings screen structure
 
+## [1.1.0-rc.1] - 2021-02-11
+
 ## [1.0.0] - 2021-02-03
-## [1.0.0-rc.1] - 2021-01-30
-### Added
 - Receive real-time push notifications
 - A list of previous notifications received
 - A detailed card of individual notifications
 - Instructions on how to use
+## [1.0.0-rc.1] - 2021-01-30
 
 [Unreleased]: https://github.com/adriancleung/pushie/compare/v1.1.0...HEAD
-[1.1.0]: https://github.com/adriancleung/pushie/compare/v1.0.0...v1.1.0
+[1.1.0]: https://github.com/adriancleung/pushie/compare/v1.1.0-rc.1...v1.1.0
+[1.1.0-rc.1]: https://github.com/adriancleung/pushie/compare/v1.0.0...v1.1.0-rc.1
 [1.0.0]: https://github.com/adriancleung/pushie/compare/v1.0.0-rc.1...v1.0.0
 [1.0.0-rc.1]: https://github.com/adriancleung/pushie/releases/tag/v1.0.0-rc.1